tweak suggestions for grid spacing and fonts
