+++ components/Resume/ResumeDescription.tsx
@@ -1,0 +1,13 @@
+export const ResumeDescription: React.FC = () => {
+  return (
+    <section>
+      <p className="text-justify pb-10">
+        I have been learning about programming and Python since I was twelve.
+        However I started to study JavaScript in 2019 and JavaScript became my
+        main language on the back-end with NodeJs and on the Front-end with
+        ReactJs and on both with NextJs. Since that I have been building
+        Full-Stack applications and trying to learn everything I can.
+      </p>
+    </section>
+  );
+};
