+++ components/Resume/ResumeEducation.tsx
@@ -1,0 +1,14 @@
+export const ResumeEducation: React.FC = () => (
+  <section className="flex flex-col gap-3">
+    <h1 className="font-mono text-2xl">EDUCATION</h1>
+    <hr />
+    <div>
+      <p className="font-bold">Tech Course</p>
+      <p>The Complete Node.js Developer Course - 2021</p>
+    </div>
+    <div>
+      <p className="font-bold">Tech Course</p>
+      <p>React The Complete Guide - 2021</p>
+    </div>
+  </section>
+);
