+++ components/Resume/ResumeLanguages.tsx
@@ -1,0 +1,11 @@
+export const ResumeLanguages = () => (
+  <section className="flex flex-col gap-3">
+    <h1 className="font-mono text-2xl">LANGUAGES</h1>
+    <hr />
+    <ul className="list-disc list-inside">
+      <li>Brazilian Portuguese (Native)</li>
+      <li>American English (Advanced)</li>
+      <li>Spanish (Basic)</li>
+    </ul>
+  </section>
+);
