+++ components/Resume/ResumeSkills.tsx
@@ -1,0 +1,13 @@
+export const ResumeSkills = () => (
+  <section className="flex flex-col gap-3">
+    <h1 className="font-mono text-2xl">SKILLS & PROFICIENCIES</h1>
+    <hr />
+    <ul className="list-disc list-inside">
+      <li>HTML5, CSS3, JavaScript, SQL, Python;</li>
+      <li>React.js and Node.Js;</li>
+      <li>Docker and Cloud development;</li>
+      <li>Teamwork, multitasking, attention to detail;</li>
+      <li>Software development;</li>
+    </ul>
+  </section>
+);
